_ext/Subject Grading/OA-Extraction/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oa-extraction"
version = "0.1.0"
description = "Handwritten question-answer extraction pipeline using Grok"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<1.0",
  "Pillow>=10.0,<12.0",
  "pydantic>=2.7,<3.0",
  "PyMuPDF>=1.24,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
oa-extract = "oa_extraction.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
